Fix user theme breakage when using LDAP auth
The LDAP authentication checks if an LDAP user already exists by
querying its database for a user with a specific e-mail address, and if
so, returns the user straight from the database.
Prior to this commit, only a select few database columns were returned,
resulting in a response to the UI containing a bunch of nil/empty
strings, which in turn made the UI save an empty string in localStorage.

Now, the GetUserByID & GetUserByEmail database queries select the same
columns, which should prevent the issue.